requirements-dev.txt
build
twine
pytest
pytest-mock
requests-mock
responses
pytest-cov
